+++ Threeyes/SDK/Scripts/Component/Manager/Simulator/System/AC_SystemAudioManagerSimulator.cs
@@ -9,17 +9,16 @@ public class AC_SystemAudioManagerSimulator : AC_SystemAudioManagerBase<AC_Syste
 	public AudioSource m_AudioSource;
 	private void Update()
 	{
-		if (!m_AudioSource.isPlaying)
-			return;
-
-		// get raw data for waveform
-		m_AudioSource.GetOutputData(rawSampleData, 0);
-		// fetch the fft
-		m_AudioSource.GetSpectrumData(fftData, 0, FFTWindow.BlackmanHarris);
-
-		// fetch the spectrum
-		m_AudioSource.GetSpectrumData(spectrumData, 0, FFTWindow.Rectangular);
-
-		OnDataChanged();
+		if (m_AudioSource.isPlaying)
+		{
+			m_AudioSource.GetOutputData(rawSampleData, 0);// get raw data for waveform			
+			m_AudioSource.GetSpectrumData(fftData, 0, FFTWindow.BlackmanHarris);// fetch the fft	
+			m_AudioSource.GetSpectrumData(spectrumData, 0, FFTWindow.Rectangular);// fetch the spectrum
+			OnDataChanged();
+		}
+		else
+		{
+			OnDataUnChanged();
+		}
 	}
 }