+++ Threeyes/SDK/Scripts/Component/Manager/Simulator/System/AC_SystemInputManagerSimulator.cs
@@ -28,6 +28,10 @@ public class AC_SystemInputManagerSimulator : AC_SystemInputManagerBase<AC_Syste
 			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.Right, 1, curCursorLocation, 0, isMouseButtonDown: true, timestamp: curTimestamp));
 		if (Input.GetMouseButtonDown(2))
 			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.Middle, 1, curCursorLocation, 0, isMouseButtonDown: true, timestamp: curTimestamp));
+		if (Input.GetMouseButtonDown(3))
+			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.XButton1, 1, curCursorLocation, 0, isMouseButtonDown: true, timestamp: curTimestamp));
+		if (Input.GetMouseButtonDown(4))
+			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.XButton2, 1, curCursorLocation, 0, isMouseButtonDown: true, timestamp: curTimestamp));
 
 		if (Input.GetMouseButtonUp(0))
 			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.Left, 1, curCursorLocation, 0, isMouseButtonUp: true, timestamp: curTimestamp));
@@ -35,6 +39,10 @@ public class AC_SystemInputManagerSimulator : AC_SystemInputManagerBase<AC_Syste
 			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.Right, 1, curCursorLocation, 0, isMouseButtonUp: true, timestamp: curTimestamp));
 		if (Input.GetMouseButtonUp(2))
 			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.Middle, 1, curCursorLocation, 0, isMouseButtonUp: true, timestamp: curTimestamp));
+		if (Input.GetMouseButtonUp(3))
+			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.XButton1, 1, curCursorLocation, 0, isMouseButtonUp: true, timestamp: curTimestamp));
+		if (Input.GetMouseButtonUp(4))
+			OnMouseDownUp(new AC_MouseEventExtArgs(AC_MouseButtons.XButton2, 1, curCursorLocation, 0, isMouseButtonUp: true, timestamp: curTimestamp));
 
 		if (Input.mouseScrollDelta != Vector2.zero)
 		{